+++ src/scripts/server/player-prefs/ui.ts
@@ -18,6 +18,7 @@ const LOOTS = [
 // This has to be kept in sync with ../player-prefs.ts
 const COVERAGE = [
     "bones",
+    "bookshelves",
     "clayLike",
     "crystals",
     "dirtLike",